refactor(ai): one-change-per-agent via AGENT_KEYS and AI_CONFIG
Add AGENT_KEYS and derive AgentKey in lib/ai/contracts/agent.ts.

Build AI_CONFIG by iterating AGENT_KEYS.

Make AI_DEFAULTS partial. Use global AI_DEFAULT_* as fallback.

Keep per-agent env overrides only for existing agents (planner, smart, coach, scheduler, moderator).

Update registry to pre-register stubs from AGENT_KEYS.

Docs: add section 'Cómo agregar un nuevo agente' in docs/ai-schema.md.
